requirements.txt
asgiref==3.3.4
Django==3.2
django-appconf==1.0.4
django-autoslug==1.9.8
django-cities-light==3.8.1
django-discover-runner==1.0
django-filter==2.4.0
django-imagekit==4.0.2
django-jazzmin==2.4.6
django-role-permissions==3.1.1
djangorestframework==3.12.4
importlib-metadata==4.0.1
Markdown==3.3.4
pilkit==2.0
Pillow==8.2.0
progressbar2==3.53.1
python-utils==2.5.6
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
Unidecode==1.2.0
zipp==3.4.1
